sprint 5

+++ src/managers/cart.js
@@ -47,37 +47,38 @@ class Cart {
         return this.carts.find(each=>each.id===id)
     }
 
-    async update_cart(id,data) {
-        try {
-            let one = this.read_cart(id)
-            for (let prop in data) {
-                one[prop] = data[prop]
-            }
-            console.log(data)
-            let data_json = JSON.stringify(this.carts,null,2)
-            await fs.promises.writeFile(this.path,data_json)
-            console.log('updated cart: '+id)
-            return 200
-        } catch(error) {
-            console.log(error)
-            return null
-        }
-    }
 
-    async update_cart(cid, pid, x) {
+    async reserve_stock(cid, pid, x) {
             try {
             let auxCart = this.read_cart(cid);
             let auxProducts = prod_manager.read_products();
             let auxProduct = prod_manager.read_product(pid);
-            if (auxProduct.stock > x) {
-                
-                auxCart.products.push({
-                id: auxProduct.id,
-                title: auxProduct.title,
-                units: x
-                });
-            }
+            if (auxProduct.stock < x || auxProduct.stock <= 0) return null
         
+            if (!auxCart) { // si no existe el carro con ese id directamente se agregara con el contenido y regresara todo ok.
+                this.carts.push({
+                    id: auxProduct.id,
+                    products: [
+                        {pid: pid, x: x}
+                    ]
+                });
+                let data_json = JSON.stringify(this.carts, null, 2);
+                await fs.promises.writeFile(this.path, data_json);
+                return 200;
+            }
+            const index = auxProducts.findIndex(e => e.id == cid)
+            if (index == -1) return null
+
+            const index2 = this.carts[index].products.findIndex(e => e.pid == pid)
+            if (index2 == -1) {
+                this.carts[index].products.push({
+                    pid: pid,
+                    x: x
+                })
+            } else {
+                console.log(this.carts[index].products[index2])
+                this.carts[index].products[index2].x += Number(x)
+            }
             for (let index = 0; index < auxProducts.length; index++) {
                 let element = auxProducts[index];
                 if (pid === element.id) {
@@ -96,57 +97,33 @@ class Cart {
         }
 
         async delete_cart(cid, pid, x) {
-                try {
-                    let auxCart = this.read_cart(cid);
-                    let auxCartProduct = findProduct(auxCart, pid, x);
-                    let auxProducts = prod_manager.read_products();
-                    console.log(auxProducts);
-                    let auxProduct = prod_manager.read_product(pid);
-                    function findProduct(auxCart, pid, x) {
-                        let foundProduct = auxCart.products.find(product => product.id === pid);
-                        if (foundProduct) {
-                            foundProduct.units -= x;
-                        }
-                        return foundProduct;
-                        } 
-                
-                    for (let index = 0; index < auxProducts.length; index++) {
-                        let element = auxProducts[index];
-                        if (pid === element.id) {
-                        auxProducts[index].stock = element.stock + x;
-                        prod_manager.update_product(pid, element);
-                        }
-                    }
-                    let data_json = JSON.stringify(this.carts, null, 2);
-                    await fs.promises.writeFile(this.path, data_json);
-                    return 200;
-                    } catch (error) {
-                    console.log(error);
-                    return null;
-                }
+            try {
+                const cart = this.read_cart(cid);
+                if (cart==undefined) return null
+
+                const product = prod_manager.read_product(pid)
+
+                const cartIndex = cart.products.findIndex(e => e.pid == pid)
+                if (cartIndex == -1) return
+
+                if (cart.products[cartIndex].x <= 0) return
+                if (cart.products[cartIndex].x-x < 0) return
+
+                cart.products[cartIndex].x -= x
+
+                cart.products = cart.products.filter(e => e.x > 0)
+                product.stock += x
+
+                let data_json = JSON.stringify(this.carts, null, 2);
+                await fs.promises.writeFile(this.path, data_json);
+                return 200;
+
+            } catch(err) {
+                console.log(err);
+                return null;
+            }
             }
 
-    // async destroy_cart(cid,pid,x) {
-    //     try {
-    //         let carts = this.carts.find(each=>each.id===cid)
-    //         let producto = this.carts.product.find(each=>each.id===pid)
-    //         let cantidad_en_carrito = producto.stock
-    //         console.log(cantidad_en_carrito)
-    //         if (one) {
-    //             this.carts = this.carts.filter(each=>each.id!==id)
-    //             this.carts.product = this.carts.product.filter(each=>each.id!==id)
-    //             let data_json = JSON.stringify(this.carts,null,2)
-    //             await fs.promises.writeFile(this.path,data_json)
-    //             console.log('delete cart: '+id)
-    //             return 200
-    //         }
-    //         console.log('not found')
-    //         return null
-    //     } catch(error) {
-    //         console.log(error)
-    //         return null
-    //     }
-    // }
 
     async destroy_cart(id) {
         try {
@@ -166,38 +143,6 @@ class Cart {
         }
     }
 
-
-
-    // async destroy_cart(cid, pid, x) {
-    //     try {
-    //         let auxCart = this.read_cart(cid);
-    //         let auxProducts = prod_manager.read_products();
-    //         let auxProduct = prod_manager.read_product(pid);
-    //         let unitscartProduct = auxCart.products(pid).units;
-
-
-
-
-
-    //         for (let index = 0; index < auxProducts.length; index++) {
-    //             let element = auxProducts[index];
-    //             if (pid === element.id) {
-    //             auxProducts[index].stock = element.stock + x;
-    //             prod_manager.update_product(pid, element);
-    //             }
-    //         }
-
-    //         // Condition for deleting cart not defined
-    //         this.carts = this.carts.filter(each => each.id !== id);
-    //         let data_json = JSON.stringify(this.carts,null,2)
-    //             await fs.promises.writeFile(this.path,data_json)
-    //             console.log('delete cart: '+id)
-    //             return 200
-    //     } catch(error) {
-    //         console.log(error)
-    //         return null
-    //     }
-    // }
 }
 let manager = new Cart('./src/data/carts.json')
